Get Station Forecasts From USDA NWCC AWDB

Get station forecasts from the USDA National Water and Climate Center Air and Water Database REST API.

Usage

get_forecasts(aoi, elements, awdb_options = set_options(), as_sf = FALSE)

Arguments

aoi

sfc POLYGON scalar, the area of interest used for performing a spatial filter on available stations in network.

elements

character vector, abbreviations or codes for variables of interest (e.g., "SMS" for "Soil Moisture Percent"). See Details for available elements and codes.

awdb_options

an awdb_options list with additional query parameters.

as_sf

boolean scalar, whether to return the data as an sf table. Default is FALSE. Repeating the spatial data across each station element and its time series can be costly.

Value

if as_sf, an sf table, otherwise a simple data.frame. The number of rows depends on the number of stations and element parameters. Time series data are included as a list column named "forecast_values".

Details

TODO!

Examples

# get forecasts for snow water equivalent (WTEQ)
get_forecasts(bear_lake, elements = "WTEQ")
#> Error in st_filter.sf(df, aoi): dplyr is not installed: install first?