Get Stations From USDA NWCC AWDB in Area of Interest

Get stations and their coordinates from the Air and Water Database REST API that fall in area of interest.

Usage

get_stations(aoi, elements, awdb_options = set_options())

Arguments

aoi

sfc POLYGON scalar, the area of interest used for performing a spatial filter on available stations in network.

elements

character vector, abbreviations or codes for variables of interest (e.g., "SMS" for "Soil Moisture Percent"). See Details for available elements and codes.

awdb_options

an awdb_options list with additional query parameters.

Value

an sf table with station metadata.

Details

You can also subset stations using query parameters in set_options().

Examples

# get stations with WTEQ elements
get_stations(
  bear_lake,
  elements = "WTEQ"
)
#> Simple feature collection with 10 features and 14 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -111.6296 ymin: 41.68541 xmax: -111.1663 ymax: 42.4132
#> Geodetic CRS:  WGS 84
#> # A tibble: 10 × 15
#>    station_triplet station_id state_code network_code name  dco_code county_name
#>    <chr>           <chr>      <chr>      <chr>        <chr> <chr>    <chr>      
#>  1 374:UT:SNTL     374        UT         SNTL         Bug … UT       Rich       
#>  2 471:ID:SNTL     471        ID         SNTL         Emig… ID       Bear Lake  
#>  3 484:ID:SNTL     484        ID         SNTL         Fran… UT       Franklin   
#>  4 1114:UT:SNTL    1114       UT         SNTL         Gard… UT       Cache      
#>  5 493:ID:SNTL     493        ID         SNTL         Give… ID       Bear Lake  
#>  6 1115:UT:SNTL    1115       UT         SNTL         Klon… UT       Cache      
#>  7 1013:UT:SNTL    1013       UT         SNTL         Temp… UT       Cache      
#>  8 823:UT:SNTL     823        UT         SNTL         Tony… UT       Cache      
#>  9 1113:UT:SNTL    1113       UT         SNTL         Tony… UT       Cache      
#> 10 1098:UT:SNTL    1098       UT         SNTL         Usu … UT       Rich       
#> # ℹ 8 more variables: huc <chr>, elevation <dbl>, data_time_zone <dbl>,
#> #   pedon_code <chr>, shef_id <chr>, begin_date <chr>, end_date <chr>,
#> #   geometry <POINT [°]>

# get stations with WTEQ elements that are in the SNTL network
get_stations(
  bear_lake,
  elements = "WTEQ",
  awdb_options = set_options(networks = "SNTL")
)
#> Simple feature collection with 10 features and 14 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -111.6296 ymin: 41.68541 xmax: -111.1663 ymax: 42.4132
#> Geodetic CRS:  WGS 84
#> # A tibble: 10 × 15
#>    station_triplet station_id state_code network_code name  dco_code county_name
#>    <chr>           <chr>      <chr>      <chr>        <chr> <chr>    <chr>      
#>  1 374:UT:SNTL     374        UT         SNTL         Bug … UT       Rich       
#>  2 471:ID:SNTL     471        ID         SNTL         Emig… ID       Bear Lake  
#>  3 484:ID:SNTL     484        ID         SNTL         Fran… UT       Franklin   
#>  4 1114:UT:SNTL    1114       UT         SNTL         Gard… UT       Cache      
#>  5 493:ID:SNTL     493        ID         SNTL         Give… ID       Bear Lake  
#>  6 1115:UT:SNTL    1115       UT         SNTL         Klon… UT       Cache      
#>  7 1013:UT:SNTL    1013       UT         SNTL         Temp… UT       Cache      
#>  8 823:UT:SNTL     823        UT         SNTL         Tony… UT       Cache      
#>  9 1113:UT:SNTL    1113       UT         SNTL         Tony… UT       Cache      
#> 10 1098:UT:SNTL    1098       UT         SNTL         Usu … UT       Rich       
#> # ℹ 8 more variables: huc <chr>, elevation <dbl>, data_time_zone <dbl>,
#> #   pedon_code <chr>, shef_id <chr>, begin_date <chr>, end_date <chr>,
#> #   geometry <POINT [°]>